requirements.txt
Pillow==9.0.1
PySimpleGUI==4.60.1
PyPDF2==3.0.1
reportlab==3.6.8